codebase/backend/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
greenlet==3.1.1
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.1
more-itertools==8.10.0
mysqlclient==1.4.6
netifaces==0.11.0
nulltype==2.3.1
oauthlib==3.2.0
pillow==10.3.0
plaid-python==26.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.10.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.4.2
systemd-python==234
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
Werkzeug==3.0.4
WTForms==3.1.2
zipp==1.0.0
